# specs/main/generated/mock-server/api/openapi.yaml
openapi: 3.0.3
info:
  title: bl1nk Skill IDE API
  version: 0.1.0
  description: |
    The bl1nk Skill IDE API powers workspace-centric development of Claude/GPT skills,
    orchestrated agents, MCP tools, curated knowledgebases, and persistent memories.
    This specification captures the minimal REST surface for desktop and cloud clients.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.bl1nk.ai/v1
    description: Production
  - url: https://staging.bl1nk.ai/v1
    description: Staging
  - url: https://dev.bl1nk.ai/v1
    description: Internal development
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      security: []
      tags:
        - Auth
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        default:
          $ref: '#/components/responses/ServerError'
      x-openapi-router-controller: AuthController
      x-openapi-router-service: AuthService
      x-eov-operation-handler: controllers/AuthController
  /auth/login:
    post:
      security: []
      tags:
        - Auth
      summary: Authenticate with email/password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/ServerError'
      x-openapi-router-controller: AuthController
      x-openapi-router-service: AuthService
      x-eov-operation-handler: controllers/AuthController
  /auth/me:
    get:
      security:
        - bearerAuth: []
      tags:
        - Auth
      summary: Get the authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/ServerError'
      x-openapi-router-controller: AuthController
      x-openapi-router-service: AuthService
      x-eov-operation-handler: controllers/AuthController
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Payload validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ConflictError:
      description: Conflicting state or duplicate resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: cuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum:
            - owner
            - admin
            - member
        status:
          type: string
          enum:
            - active
            - invited
            - disabled
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true